pyproject.toml
[project]
name = "soccermatics-pro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "mplsoccer>=1.6.1",
    "omegaconf>=2.3.0",
    "pandas-stubs==2.3.2.250926",
    "statsmodels>=0.14.5",
    "streamlit>=1.51.0",
    "streamlit-tags>=1.2.8",
]
